+++ src/Runtime/mmdModel.ts
@@ -231,6 +231,8 @@ export class MmdModel implements IMmdModel {
 
     /**
      * Add an animation to this model
+     *
+     * If the animation is already added, it will be replaced
      * @param animation MMD animation or MMD model animation group to add
      * @param retargetingMap Model bone name to animation bone name map
      */
@@ -244,8 +246,14 @@ export class MmdModel implements IMmdModel {
         } else {
             throw new Error("animation is not MmdAnimation or MmdModelAnimationGroup or MmdCompositeAnimation. are you missing import \"babylon-mmd/esm/Runtime/Animation/mmdRuntimeModelAnimation\" or \"babylon-mmd/esm/Runtime/Animation/mmdRuntimeModelAnimationGroup\" or \"babylon-mmd/esm/Runtime/Animation/mmdCompositeRuntimeModelAnimation\"?");
         }
-        this._animationIndexMap.set(animation.name, this._animations.length);
-        this._animations.push(runtimeAnimation);
+
+        const index = this._animationIndexMap.get(animation.name);
+        if (index !== undefined) {
+            this._animations[index] = runtimeAnimation;
+        } else {
+            this._animationIndexMap.set(animation.name, this._animations.length);
+            this._animations.push(runtimeAnimation);
+        }
     }
 
     /**